fix duplicated record from database; fix delete service; add redirect page; add length validation to form.

+++ src/main/java/com/walletERP/controller/service/delete/DeleteCustomersBean.java
@@ -2,21 +2,19 @@ package com.walletERP.controller.service.delete;
 
 import com.walletERP.model.dao.CustomerDAO;
 import com.walletERP.model.entity.Customer;
-import org.springframework.web.context.annotation.SessionScope;
 
 import javax.annotation.ManagedBean;
 import javax.enterprise.context.RequestScoped;
 import javax.faces.application.FacesMessage;
-import javax.faces.context.ExternalContext;
 import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import java.io.Serializable;
 
 @ManagedBean
-@SessionScope
+@RequestScoped
 public class DeleteCustomersBean implements Serializable {
-    private static final long serialVersionUID = 4739757182506178525L;
-    private final Customer customer;
+    private static final long serialVersionUID = 3528969511017885937L;
+    private Customer customer;
     private final CustomerDAO customerDAO;
 
     @Inject
@@ -25,13 +23,24 @@ public class DeleteCustomersBean implements Serializable {
         this.customerDAO = customerDAO;
     }
 
-    public void deleteCustomer(){
-        ExternalContext ec = FacesContext.getCurrentInstance().getExternalContext();
-        String customerId = ec.getRequestParameterMap().get("deleteForm:customerIdDelete");
-        this.customer.setCustomerID(Long.valueOf(customerId));
-        this.customerDAO.deleteCustomerByID(customer);
+    public Customer getCustomer() {
+        return customer;
+    }
+
+    public void setCustomer(Customer customer) {
+        this.customer = customer;
+    }
+
+    public String deleteCustomer() {
+        String customerID = FacesContext.getCurrentInstance().getExternalContext().getRequestParameterMap().get("customerIdDelete");
+        if (customerID != null) {
+            this.customer.setCustomerID(Long.valueOf(customerID));
+            this.customerDAO.deleteCustomerByID(customer);
+        }
 
         FacesMessage msg = new FacesMessage("Successful", "User delete successfully.");
         FacesContext.getCurrentInstance().addMessage(null, msg);
+
+        return "refresh?faces-redirect = true";
     }
 }
